refactor: changes after review
