+++ src/types/loyaltyPass.ts
@@ -1,6 +1,11 @@
 import type { QuestDetails } from '@/types/questDetails';
 import { type RootNode } from 'node_modules/@strapi/blocks-react-renderer/dist/BlocksRenderer';
-import type { CampaignData, StrapiMediaData } from './strapi';
+import type {
+  CampaignData,
+  StrapiMediaData,
+  TaskType,
+  TaskWidgetInformationData,
+} from './strapi';
 
 // PDA Type
 export interface Reward {
@@ -114,39 +119,6 @@ export interface ParticipantChain {
   avatarUrl: string;
   label: string;
   id: number;
-}
-
-export enum TaskType {
-  Bridge = 'Bridge',
-  Swap = 'Swap',
-  Deposit = 'Deposit',
-  OnChain = 'On-chain',
-  OffChain = 'Off-chain',
-  Zap = 'Zap',
-}
-
-export interface TaskWidgetInformationChainData {
-  chainId: string;
-  chainKey: string;
-}
-
-export interface TaskWidgetInformationTokenData {
-  tokenAddress: string;
-  tokenSymbol: string;
-}
-
-export interface TaskWidgetInformationWalletData {
-  walletAddress: string;
-  chainType: string;
-}
-
-export interface TaskWidgetInformationData {
-  sourceChain?: TaskWidgetInformationChainData;
-  sourceToken?: TaskWidgetInformationTokenData;
-  destinationChain?: TaskWidgetInformationChainData;
-  destinationToken?: TaskWidgetInformationTokenData;
-  toAddress?: TaskWidgetInformationWalletData;
-  fromAmount?: string;
 }
 
 export interface TaskVerification {
